requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
firebase-admin==6.2.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
gunicorn==21.2.0
